webapi/tct-webaudio-w3c-tests/webaudio/audionode_numberOfOutputs_type.html

<!DOCTYPE html>
<!--
Copyright (c) 2012 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Ge, Qing <qingx.ge@intel.com>

-->

<html>
  <head>
    <title>WebAudio Test: audionode_numberOfOutputs_type</title>
    <link rel="author" title="Intel" href="http://www.intel.com" />
    <link rel="help" href="https://webaudio.github.io/web-audio-api/#the-audionode-interface" />
    <meta name="flags" content="" />
    <meta name="assert" content="Check if AudioNode.numberOfOutputs is of type number" />
    <script src="../resources/testharness.js"></script>
    <script src="../resources/testharnessreport.js"></script>
    <script src="./support/webaudio.js"></script>
  </head>
  <body>
    <div id=log></div>
    <script type="text/javascript">
        test(function () {
            var context = new AudioContext();
            if(!("createBufferSource" in context)){
                assert_true(false, "createBufferSource method exist AudioContext interface");
            }else{
                var sourceNode = context.createBufferSource();
                if(!("numberOfOutputs" in sourceNode)){
                    assert_true(false, "numberOfOutputs attribute exist AudioNode interface");
                }else{
                    assert_true(typeof sourceNode.numberOfOutputs == "number", "Check if AudioNode.numberOfOutputs is of type number");
                }
            }
        }, document.title);
    </script>
  </body>
</html>
